controller delete update

+++ backend/controllers/product.controller.js
@@ -23,6 +23,13 @@ export const getProducts = async(req, res) => {
 export const deleteProduct = async(req, res) => {
     const {id} = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({
+            success: false,
+            message: "Product Not Found"
+        })
+    }
+
     try{
         await Product.findByIdAndDelete(id);
         return res.status(200).json({
@@ -31,9 +38,9 @@ export const deleteProduct = async(req, res) => {
         });
     } catch (error){
         console.error("Product not found", error.message)
-        return res.status(404).json({
+        return res.status(500).json({
             success: false,
-            message: "Product not found"
+            message: "Server Error"
         })
     }
 };
